+++ chevah/empirical/tests/normal/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) 2012 Adi Roiban.
+# See LICENSE for details.
+"""
+Unit tests for empirical package.
+"""